MDEV-23399 fixup: Fix a hang in innodb_flush_sync=ON mode
buf_flush_do_batch(): Always ensure that all log has been written.
We used to limit the amount written when performing the 'furious flushing'
for a log checkpoint in innodb_flush_sync=ON mode. However, it is
possible that a page that we must flush due to small oldest_modification
also carries a FIL_PAGE_LSN that is much later than the checkpoint
target LSN. Therefore, we must flush the entire log at the start of
a batch.